Refactor(changhoon/signup) : #28
	1. 함수 분리
	2. 사용하지 않는 test파일 삭제
