+++ app/src/apis/signup/index.js
@@ -5,6 +5,7 @@ const express = require('express');
 const router = express.Router();
 const ctrl = require('./signup.ctrl');
 
+// 회원가입 API
 router.post('/signup', ctrl.process.signup);
 
 module.exports = router;
