--- app/src/apis/test.js
@@ -1,1 +1,0 @@
-// 임시 파일
