--- app/src/models/services/test.js
@@ -1,1 +1,0 @@
-// 임시 파일
