--- app/src/public/js/test.js
@@ -1,1 +1,0 @@
-// 임시 파일
